Development

+++ uu_shoppinglist_maing01-server/app/abl/list-abl.js
@@ -98,7 +98,7 @@ class ListAbl {
     return dtoOut;
   }
 
-  async get(dtoIn, awid) {
+  async get(dtoIn, awid, session) {
     let uuAppErrorMap = {};
 
     // validates dtoIn
@@ -111,15 +111,21 @@ class ListAbl {
       Errors.Get.InvalidDtoIn
     );
 
+    const uuIdentity = session.getIdentity().getUuIdentity();
+
     //fetch from db
-    const dtoOut = await this.dao.get(awid, dtoIn.id);
+    const result = await this.dao.get(awid, dtoIn.id);
+
+    if (result.creatorUuId !== uuIdentity && !result.members.find((member) => member.uuId === uuIdentity)) {
+      throw new Errors.Get.UserNotAuthorized({ uuAppErrorMap });
+    }
 
     // prepare and return dtoOut
-    dtoOut.uuAppErrorMap = uuAppErrorMap;
+    const dtoOut = { ...result, uuAppErrorMap };
     return dtoOut;
   }
 
-  async delete(dtoIn, awid) {
+  async delete(dtoIn, awid, session) {
     let uuAppErrorMap = {};
 
     // validates dtoIn
@@ -132,6 +138,14 @@ class ListAbl {
       Errors.Delete.InvalidDtoIn
     );
 
+    const list = await this.dao.get(awid, dtoIn.id);
+
+    const uuIdentity = session.getIdentity().getUuIdentity();
+
+    if (list.creatorUuId !== uuIdentity) {
+      throw new Errors.Delete.UserNotAuthorized({ uuAppErrorMap });
+    }
+
     //delete from db
     const result = await this.dao.delete(awid, dtoIn.id);
 
@@ -140,7 +154,7 @@ class ListAbl {
     return dtoOut;
   }
 
-  async deleteItem(dtoIn, awid) {
+  async deleteItem(dtoIn, awid, session) {
     let uuAppErrorMap = {};
 
     // validates dtoIn
@@ -153,6 +167,14 @@ class ListAbl {
       Errors.DeleteItem.InvalidDtoIn
     );
 
+    const list = await this.dao.get(awid, dtoIn.listId);
+
+    const uuIdentity = session.getIdentity().getUuIdentity();
+
+    if (list.creatorUuId !== uuIdentity && !list.members.find((member) => member.uuId === uuIdentity)) {
+      throw new Errors.DeleteItem.UserNotAuthorized({ uuAppErrorMap });
+    }
+
     dtoIn.itemId = new ObjectId(dtoIn.itemId);
 
     //delete from list in db
@@ -163,7 +185,7 @@ class ListAbl {
     return dtoOut;
   }
 
-  async deleteMember(dtoIn, awid) {
+  async deleteMember(dtoIn, awid, session) {
     let uuAppErrorMap = {};
 
     // validates dtoIn
@@ -176,6 +198,17 @@ class ListAbl {
       Errors.DeleteMember.InvalidDtoIn
     );
 
+    const list = await this.dao.get(awid, dtoIn.listId);
+
+    const uuIdentity = session.getIdentity().getUuIdentity();
+
+    if (
+      list.creatorUuId !== uuIdentity &&
+      (!list.members.find((member) => member.uuId === uuIdentity) || dtoIn.memberUuId !== uuIdentity)
+    ) {
+      throw new Errors.DeleteMember.UserNotAuthorized({ uuAppErrorMap });
+    }
+
     //delete from list in db
     const result = await this.dao.deleteMember(awid, dtoIn.listId, dtoIn.memberUuId);
 
@@ -184,7 +217,7 @@ class ListAbl {
     return dtoOut;
   }
 
-  async update(dtoIn, awid) {
+  async update(dtoIn, awid, session) {
     let uuAppErrorMap = {};
 
     // validates dtoIn
@@ -196,6 +229,14 @@ class ListAbl {
       Warnings.Update.UnsupportedKeys.code,
       Errors.Update.InvalidDtoIn
     );
+
+    const listBeforeUpdate = await this.dao.get(awid, dtoIn.id);
+
+    const uuIdentity = session.getIdentity().getUuIdentity();
+
+    if (listBeforeUpdate.creatorUuId !== uuIdentity) {
+      throw new Errors.Update.UserNotAuthorized({ uuAppErrorMap });
+    }
 
     let list = {};
 
@@ -210,7 +251,7 @@ class ListAbl {
     return dtoOut;
   }
 
-  async updateItem(dtoIn, awid) {
+  async updateItem(dtoIn, awid, session) {
     let uuAppErrorMap = {};
 
     // validates dtoIn
@@ -223,6 +264,14 @@ class ListAbl {
       Errors.UpdateItem.InvalidDtoIn
     );
 
+    const list = await this.dao.get(awid, dtoIn.listId);
+
+    const uuIdentity = session.getIdentity().getUuIdentity();
+
+    if (list.creatorUuId !== uuIdentity && !list.members.find((member) => member.uuId === uuIdentity)) {
+      throw new Errors.UpdateItem.UserNotAuthorized({ uuAppErrorMap });
+    }
+
     dtoIn.itemId = new ObjectId(dtoIn.itemId);
 
     //update in list in db
@@ -233,7 +282,7 @@ class ListAbl {
     return dtoOut;
   }
 
-  async addMember(dtoIn, awid) {
+  async addMember(dtoIn, awid, session) {
     let uuAppErrorMap = {};
 
     // validates dtoIn
@@ -246,6 +295,14 @@ class ListAbl {
       Errors.AddMember.InvalidDtoIn
     );
 
+    const list = await this.dao.get(awid, dtoIn.listId);
+
+    const uuIdentity = session.getIdentity().getUuIdentity();
+
+    if (list.creatorUuId !== uuIdentity) {
+      throw new Errors.AddMember.UserNotAuthorized({ uuAppErrorMap });
+    }
+
     const member = { uuId: dtoIn.memberUuId, name: dtoIn.memberName };
 
     //add into list in db
@@ -256,7 +313,7 @@ class ListAbl {
     return dtoOut;
   }
 
-  async addItem(dtoIn, awid) {
+  async addItem(dtoIn, awid, session) {
     let uuAppErrorMap = {};
 
     // validates dtoIn
@@ -268,6 +325,14 @@ class ListAbl {
       Warnings.AddItem.UnsupportedKeys.code,
       Errors.AddItem.InvalidDtoIn
     );
+
+    const list = await this.dao.get(awid, dtoIn.listId);
+
+    const uuIdentity = session.getIdentity().getUuIdentity();
+
+    if (list.creatorUuId !== uuIdentity && !list.members.find((member) => member.uuId === uuIdentity)) {
+      throw new Errors.AddItem.UserNotAuthorized({ uuAppErrorMap });
+    }
 
     const item = { name: dtoIn.itemName, solved: dtoIn.solved, id: new ObjectId() };
